Adds key input support
This implements basic key board input for the chip8's hex keyboard. The
hex keypad directly corresponds to the ASCII for example, pressing 'a'
while send 'a' to the  emulator. ASCII key codes outside of 0-9 and a-f
do nothing. A future commit will add support for key maps.

src/ui.c:
* Adds key press handling code
* Adds hard coded key map

src/chip8.c:
* Fixes a bug where the keyMutex would not be unlocked
* Modifies the chip8 initializer to set the "chip" field on the ui
struct to the current chip8 object